Implemented image loader and converter, as well as Basis Universal compressor and transcoder, added YAML serialization of textures and materials, created serialization wizards, introduced Framebuffers, improved UI, made a Stream interface

+++ Sierra/src/Rendering/SwapchainEvents.cpp
@@ -9,7 +9,7 @@ namespace Sierra
 
     /* --- CONSTRUCTORS --- */
 
-    SwapchainResizeEvent::SwapchainResizeEvent(const uint32 width, const uint32 height, const uint32 scaling) noexcept
+    SwapchainResizeEvent::SwapchainResizeEvent(const uint32 width, const uint32 height, const float32 scaling) noexcept
         : scaling(scaling), width(width), height(height)
     {
